+++ apps/cloud/src/app/index.tsx
@@ -1,11 +1,12 @@
-import { ErrorBoundary } from 'react-error-boundary';
 import { ErrorFallback, ReduxProvider, StylesProvider } from './components';
-import { KUICombineProviders } from '@kleeen/core-react';
-import { LocalizationContextProvider } from '@kleeen/react/hooks';
+import { LocalizationContextProvider, PreviewPanelLayoutProvider } from '@kleeen/react/hooks';
 import { MessageShape, ToastNotification } from '@kleeen/react/atomic-elements';
+
+import AppContainer from './app';
+import { ErrorBoundary } from 'react-error-boundary';
+import { KUICombineProviders } from '@kleeen/core-react';
 import { SnackbarProvider } from 'notistack';
 import { useState } from 'react';
-import AppContainer from './app';
 
 export function App(): JSX.Element {
   const [errorInfo, setErrorInfo] = useState<{ componentStack: string }>(null);
@@ -37,7 +38,9 @@ export function App(): JSX.Element {
               content={(key, message) => <ToastNotification id={key} message={message as MessageShape} />}
               maxSnack={10}
             >
-              <AppContainer />
+              <PreviewPanelLayoutProvider>
+                <AppContainer />
+              </PreviewPanelLayoutProvider>
             </SnackbarProvider>
           </KUICombineProviders>
         </LocalizationContextProvider>