+++ libs/react/components/src/lib/context-menu/context-menu.tsx
@@ -1,68 +1,42 @@
-import { ContextMenuProps, MenuItemType } from './context-menu.model';
-import { FilterTypes, Widget } from '@kleeen/types';
-import { Menu, MenuGroupItem, MenuGroupName, MenuTitle } from './contextual-menu.style';
+import { ContextMenuProps, ContextMenuSectionItem, FormattedContextDataPoint } from './context-menu.model';
+import { Menu, MenuTitle } from './contextual-menu.style';
+import { ReactNode, useEffect, useRef, useState } from 'react';
 import {
-  PreviewItem,
-  useCrossLinkingItems,
-  useFilterItems,
-  useHoverIntent,
-  usePreviewItems,
-  usePreviewPanel,
-  useTextFormatter,
-  useTheme,
-} from '@kleeen/react/hooks';
-import { isNilOrEmpty, roleAccessKeyTag } from '@kleeen/common/utils';
-import { useEffect, useRef } from 'react';
+  useCrossLinkingSections,
+  useDataPointsWithFormattedValue,
+  useFilterSections,
+  usePreviewSections,
+} from './hooks';
+import { useHoverIntent, useTheme } from '@kleeen/react/hooks';
 
-import { AccessControl } from '@kleeen/core-react';
-import Tooltip from '@material-ui/core/Tooltip';
-import { Translate } from '@kleeen/core-react';
-import { WidgetScope } from '@kleeen/widgets';
+import { AggregationType } from '@kleeen/types';
+import { ContextMenuItemView } from './context-menu-item';
+import { ContextMenuSection } from './context-menu-section';
+import { DEFAULT_TRANSFORMATION_KEY_TO_USE } from './utils';
+import { isNilOrEmpty } from '@kleeen/common/utils';
+import { isSingleCardinalityTransformation } from '@kleeen/frontend/utils';
+import { path } from 'ramda';
 
-const hidePermission = 'HIDE';
-
-export const KsContextMenu = ({
-  anchorEl,
-  autoClose,
-  attr,
-  cell,
-  displayColumnAttribute,
-  handleClose,
-  row,
-}: ContextMenuProps): JSX.Element => {
-  const { themeClass } = useTheme();
-  const timerRef = useRef(null);
+export function KsContextMenu({ anchorEl, autoClose, dataPoints, handleClose }: ContextMenuProps) {
+  const formattedDataPoints = useDataPointsWithFormattedValue({ dataPoints });
   const { ref } = useHoverIntent<HTMLUListElement>({
     delayOnEnter: 0,
     onMouseEnterFn: clearTimeOut,
     onMouseLeaveFn: handleClose,
   });
+  const timerRef = useRef(null);
+  const [dataPointsToShow, setDataPointsToShow] = useState<FormattedContextDataPoint[]>([]);
+  const [menuSections, setMenuSections] = useState<ContextMenuSectionItem[]>([]);
+  const [menuTitle, setMenuTitle] = useState<ReactNode>();
+  const { themeClass } = useTheme();
 
-  const crossLinkItems = useCrossLinkingItems({
-    handleClose,
-    attr,
-    cell,
-  });
-  const [formatter] = useTextFormatter({
-    format: attr?.format,
-    formatType: attr?.formatType,
-    transformation: attr?.aggregation,
-  });
-  const formattedValue = formatter(cell?.displayValue);
-  const filtersMenuItems = useFilterItems({
-    attr,
-    cell: { ...cell, formattedValue },
+  const crossLinkSection = useCrossLinkingSections({ dataPointsToShow, handleClose });
+  const filterSections = useFilterSections({ dataPointsToShow, handleClose });
+  const previewSections = usePreviewSections({
+    dataPoints: formattedDataPoints,
+    dataPointsToShow,
     handleClose,
   });
-
-  const previewItems = usePreviewItems({
-    attr,
-    cell: { ...cell, formattedValue },
-    displayColumnAttribute,
-    row,
-  });
-  // TODO: Move these inside a useEffect
-  const previewSectionAndItems = getPreviewItemsAsMenuItems({ handleClose, previewItems });
 
   useEffect(() => {
     if (autoClose) {
@@ -76,6 +50,62 @@ export const KsContextMenu = ({
     }
   }, []);
 
+  useEffect(() => {
+    if (isNilOrEmpty(dataPoints)) {
+      setDataPointsToShow([]);
+      return;
+    }
+
+    const tempFilteredDataPoints = formattedDataPoints
+      .filter(({ ignoreInContextMenu = false }) => !ignoreInContextMenu)
+      .sort((a, b) => {
+        const aCardinalityWeight = Number(
+          isSingleCardinalityTransformation(a.attribute.aggregation as AggregationType),
+        );
+        const bCardinalityWeight = Number(
+          isSingleCardinalityTransformation(b.attribute.aggregation as AggregationType),
+        );
+        return bCardinalityWeight - aCardinalityWeight;
+      });
+
+    setDataPointsToShow(tempFilteredDataPoints);
+  }, [formattedDataPoints?.length]);
+
+  useEffect(() => {
+    if (isNilOrEmpty(dataPointsToShow)) {
+      return;
+    }
+
+    const [firstDataPoint] = dataPointsToShow;
+    // TODO @cafe THIS MUST BE REMOVED ONCE WE GET RID OF THE AGGREGATION VS TRANSFORMATION DILEMMA.
+    const { transformationKeyToUse = DEFAULT_TRANSFORMATION_KEY_TO_USE } = firstDataPoint;
+    const attributeTransformation = path<AggregationType>([transformationKeyToUse], firstDataPoint.attribute);
+    const isSingleCardinality = isSingleCardinalityTransformation(attributeTransformation);
+
+    const menuTitle = `${firstDataPoint.formattedValue} ${
+      !isSingleCardinality ? firstDataPoint.attribute.name : ''
+    }`;
+    setMenuTitle(menuTitle);
+  }, [dataPointsToShow?.length]);
+
+  useEffect(() => {
+    const sectionItems = [];
+
+    if (!isNilOrEmpty(crossLinkSection)) {
+      sectionItems.push(...crossLinkSection);
+    }
+
+    if (!isNilOrEmpty(filterSections)) {
+      sectionItems.push(...filterSections);
+    }
+
+    if (!isNilOrEmpty(previewSections)) {
+      sectionItems.push(...previewSections);
+    }
+
+    setMenuSections(sectionItems);
+  }, [crossLinkSection?.length, filterSections?.length, previewSections?.length]);
+
   function clearTimeOut() {
     if (timerRef.current) {
       clearTimeout(timerRef.current);
@@ -83,161 +113,34 @@ export const KsContextMenu = ({
     }
   }
 
-  const filterInMenuItems = filtersMenuItems.length
-    ? filtersMenuItems.filter((item) => item.filterType === FilterTypes.in)
-    : [];
-  const filterOutMenuItems = filtersMenuItems.length
-    ? filtersMenuItems.filter((item) => item.filterType === FilterTypes.out)
-    : [];
-
-  const prepareCrossLinkingSectionAndItems = crossLinkItems.length
-    ? [
-        {
-          type: MenuItemType.Section,
-          label: (
-            <Translate
-              id="app.contextMenu.goTo.value"
-              type="html"
-              values={{ value: formattedValue as string }}
-            />
-          ),
-        },
-        ...crossLinkItems,
-      ]
-    : [];
-  const prepareFilterInSectionAndItems = filterInMenuItems.length
-    ? [
-        {
-          type: MenuItemType.Section,
-          label: (
-            <Translate id="app.contextMenu.filter.addAndApply" type="html" values={{ value: attr?.label }} />
-          ),
-        },
-        ...filterInMenuItems,
-      ]
-    : [];
-  const prepareFilterOutSectionAndItems = filterOutMenuItems.length
-    ? [
-        {
-          type: MenuItemType.Section,
-          label: (
-            <Translate
-              id="app.contextMenu.filter.removeAndApply"
-              type="html"
-              values={{ value: attr?.label }}
-            />
-          ),
-        },
-        ...filterOutMenuItems,
-      ]
-    : [];
-
-  const menuItems = [
-    ...prepareCrossLinkingSectionAndItems,
-    ...prepareFilterInSectionAndItems,
-    ...prepareFilterOutSectionAndItems,
-    ...previewSectionAndItems,
-  ];
-
-  if (isNilOrEmpty(menuItems)) return null;
+  if (isNilOrEmpty(menuSections)) return null;
 
   return (
-    <>
-      <Menu
-        id="context-menu"
-        className={themeClass}
-        anchorEl={anchorEl}
-        anchorOrigin={{
-          horizontal: 'center',
-          vertical: 'center',
-        }}
-        open={Boolean(anchorEl)}
-        onClose={handleClose}
-        getContentAnchorEl={null}
-        MenuListProps={{ ref }}
-      >
-        <MenuTitle>{formattedValue}</MenuTitle>
-        {menuItems.map((item, i) => {
-          if (item.type === MenuItemType.Section)
-            return (
-              <MenuGroupName key={i}>
-                <Tooltip title={item.label} placement="top">
-                  <span className="menu-item-text">{item.label}</span>
-                </Tooltip>
-              </MenuGroupName>
-            );
+    <Menu
+      anchorEl={anchorEl}
+      anchorOrigin={{
+        horizontal: 'center',
+        vertical: 'center',
+      }}
+      className={themeClass}
+      getContentAnchorEl={null}
+      id="context-menu"
+      onClose={handleClose}
+      open={Boolean(anchorEl)}
+      MenuListProps={{ ref }}
+    >
+      <MenuTitle>{menuTitle}</MenuTitle>
+      {menuSections.map((section) => {
+        const { menuItems } = section;
 
-          const itemAccessKey = roleAccessKeyTag(item.roleAccessKey);
-
-          return (
-            <AccessControl id={itemAccessKey} key={item.key}>
-              {({ permission }) => (
-                <MenuGroupItem
-                  disabled={permission === hidePermission}
-                  key={item.key}
-                  onClick={item.handleClick(item)}
-                  className={`${i % 2 ? 'odd-stripe' : 'even-stripe'}`}
-                >
-                  <Tooltip title={item.label} placement="top">
-                    <span className="menu-item-text">{item.label}</span>
-                  </Tooltip>
-                </MenuGroupItem>
-              )}
-            </AccessControl>
-          );
-        })}
-      </Menu>
-    </>
+        return (
+          <ContextMenuSection section={section}>
+            {menuItems.map((item, index) => {
+              return <ContextMenuItemView index={index} item={item} />;
+            })}
+          </ContextMenuSection>
+        );
+      })}
+    </Menu>
   );
-};
-
-//#region Private Members
-interface GetPreviewItemsAsMenuItemsParams {
-  handleClose: () => void;
-  previewItems: PreviewItem[];
 }
-
-function getPreviewItemsAsMenuItems({ handleClose, previewItems }: GetPreviewItemsAsMenuItemsParams) {
-  const previewContext = usePreviewPanel();
-
-  if (isNilOrEmpty(previewItems)) {
-    return [];
-  }
-
-  return [
-    {
-      type: MenuItemType.Section,
-      label: <Translate id="app.contextMenu.preview" type="html" />,
-    },
-    ...previewItems.map(({ entity, filteredBy, scope, value, widgets }) => {
-      return scope === WidgetScope.Single
-        ? {
-            label: <Translate id="app.contextMenu.preview.single" type="html" values={{ value }} />,
-            handleClick: () => () => {
-              handleClose();
-              previewContext.setPreviewWidgets(widgets as Widget[]);
-              previewContext.openPreviewPanel();
-            },
-            key: `preview.single`,
-            roleAccessKey: `preview.single`,
-          }
-        : {
-            label: (
-              <Translate
-                id="app.contextMenu.preview.collection"
-                type="html"
-                values={{ entity, filteredBy }}
-              />
-            ),
-            handleClick: () => () => {
-              handleClose();
-              previewContext.setPreviewWidgets(widgets as Widget[]);
-              previewContext.openPreviewPanel();
-            },
-            key: `preview.collection`,
-            roleAccessKey: `preview.collection`,
-          };
-    }),
-  ];
-}
-//#endregion