+++ libs/widgets/src/lib/types/widgets-by-entity.model.ts
@@ -65,6 +65,5 @@ export type WidgetsByEntity = WidgetsByEntityMap<LibraryWidget>;
 
 export interface WidgetByEntityBaseParam {
   entityId: number | string;
-  filters?: Filters;
   scope: WidgetScope;
 }